+++ src/component/MyPageType/MyPageLikes.tsx
@@ -5,7 +5,8 @@ import {
 } from '@tanstack/react-query'
 import { PlayList } from '@/component'
 import { useAuthStore } from '@/store/useAuthStore'
-import { DeleteLikeList, LikedPlayList } from '@/apis/userInfoApi'
+import { LikedPlayList } from '@/apis/like/fetchProfileLikePlayList'
+import { DeleteLikeList } from '@/apis/like/deleteProfileListPlaylist'
 import * as S from './MyPageLikes.styled'
 import { useInfiniteScroll, useToast } from '@/hooks'
 import { useRef } from 'react'
