+++ src/pages/mypage/index.tsx
@@ -4,7 +4,7 @@ import * as S from '@/pages/mypage/Mypage.styled'
 import { Profile, Loading, Button } from '@/component'
 import { ROUTER_PATH } from '@/constants/constant'
 import { useQuery } from '@tanstack/react-query'
-import { userInfoGet } from '@/apis/userInfoApi'
+import { userInfoGet } from '@/apis/user/fetchUserDataType'
 import { useAuthStore } from '@/store/useAuthStore'
 import { useState } from 'react'
 import MyPageLikes from '@/component/MyPageType/MyPageLikes'
