+++ src/pages/profileEdit/index.tsx
@@ -4,7 +4,7 @@ import { Profile, Input, Button } from '@/component'
 import img from '@/assets/img/profile/default_profile.png'
 import { useForm } from 'react-hook-form'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { UserProfileEdit } from '@/apis/userInfoApi'
+import { UserProfileEdit } from '@/apis/profileEdit/index'
 import { useAuthStore } from '@/store/useAuthStore'
 import { CiCirclePlus } from 'react-icons/ci'
 import { useUserInfo, useToast } from '@/hooks'
@@ -119,7 +119,6 @@ export function ProfileEdit() {
     navigate('/mypage')
   }
 
-  console.log(image?.[0])
   return (
     <>
       <S.ContentContainer onSubmit={handleSubmit(onSubmit)}>
